Initial commit
Motivation:

h3spec is very useful to validate that an quic / h3 implementation is correctly handling validations / errors. At the moment there is no automated way to do this testing during a build of a maven project.

Modifications:

Add code for a custom maven plugin which will execute the h3spec test tool during the build and fails the build if an error is detected.

Result:

Be able to execute the h3spec tool during testing