Add option to wait for a ready message before running h3spec test
Motivation:
The millisecond delay might not be sufficiently reliable in some environments, and can also be wasteful in how long builds will wait to ensure reliable execution.

Modification:
Make it possible for the test to snoop on the System.out stream and look for a readiness message instead, as the signal to start running the test.

Result:
This should produce more reliable build executions.